improved test for env testing
