+++ runtime/integration-tests/Cargo.toml
@@ -9,7 +9,7 @@ repository = "https://github.com/centrifuge/centrifuge-chain"
 
 
 [dependencies]
-fudge = { git = "https://github.com/centrifuge/fudge", tag = "v0.0.3-polkadot-v0.9.17"}
+fudge = { git = "https://github.com/centrifuge/fudge", tag = "v0.0.4-polkadot-v0.9.17"}
 tokio = { version = "1.15", features = ["macros"] }
 tracing = "0.1"
 tracing-subscriber = "0.2"
