+++ runtime/integration-tests/src/pools/mod.rs
@@ -10,5 +10,6 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
+mod env;
 mod pool_starts;
 mod utils;
