+++ runtime/integration-tests/src/pools/pool_starts.rs
@@ -21,52 +21,4 @@ use sp_runtime::Storage;
 use tokio::runtime::Handle;
 
 #[tokio::test]
-async fn env_works() {
-	let manager = env::task_manager(Handle::current());
-	let mut env = env::test_env_default(manager.spawn_handle());
-
-	let num_blocks = 10;
-	let block_before = env
-		.with_state(Chain::Para(PARA_ID), || {
-			frame_system::Pallet::<Runtime>::block_number()
-		})
-		.unwrap();
-
-	env::pass_n(num_blocks, &mut env).unwrap();
-
-	let block_after = env
-		.with_state(Chain::Para(PARA_ID), || {
-			frame_system::Pallet::<Runtime>::block_number()
-		})
-		.unwrap();
-
-	assert_eq!(block_before + num_blocks as u32, block_after)
-}
-
-#[tokio::test]
-async fn create_pool() {
-	let manager = env::task_manager(Handle::current());
-	let mut genesis = Storage::default();
-	env::default_balances::<Runtime>(&mut genesis);
-	//let mut env = env::test_env_with_centrifuge_storage(manager.spawn_handle(), genesis);
-	let mut env = env::test_env_default(manager.spawn_handle());
-
-	let to: centrifuge::Address = centrifuge::Address::Id(Keyring::Bob.to_account_id().into());
-	let xt = ext_centrifuge(
-		&env,
-		Keyring::Alice,
-		centrifuge::Call::Balances(BalancesCall::transfer {
-			dest: to,
-			value: 100 * centrifuge::CFG,
-		}),
-	);
-	env.append_extrinsic(Chain::Para(PARA_ID), xt.encode())
-		.unwrap();
-	env.evolve().unwrap();
-
-	let _info = env
-		.with_state(Chain::Para(PARA_ID), || {
-			frame_system::Pallet::<Runtime>::account(Keyring::Alice.to_account_id())
-		})
-		.unwrap();
-}
+async fn create_pool() {}
